clang-tidy: various fixes

+++ src/sonymn_int.cpp
@@ -1034,8 +1034,8 @@ std::ostream& SonyMakerNote::printWBShiftABGMPrecise(std::ostream& os, const Val
   }
   std::ios::fmtflags f(os.flags());
 
-  const auto temp0 = static_cast<double>(value.toInt64(0)) / double(1000.0);
-  const auto temp1 = static_cast<double>(value.toInt64(1)) / double(1000.0);
+  const auto temp0 = static_cast<double>(value.toInt64(0)) / (1000.0);
+  const auto temp1 = static_cast<double>(value.toInt64(1)) / (1000.0);
 
   os << "A/B: ";
   if (temp0 == 0) {
@@ -1067,7 +1067,7 @@ std::ostream& SonyMakerNote::printExposureStandardAdjustment(std::ostream& os, c
   std::ios::fmtflags f(os.flags());
 
   const auto [r, s] = value.toRational();
-  os << std::fixed << std::setprecision(1) << (double(r) / double(s));
+  os << std::fixed << std::setprecision(1) << (static_cast<double>(r) / static_cast<double>(s));
   os.flags(f);
 
   return os;
